azure_storage_blobs

Module prelude

Re-exports

pub use crate::blob::Blob;

pub use crate::blob::BlobBlockType;

pub use crate::blob::BlockList;

pub use crate::blob::BlockListType;

pub use crate::container::PublicAccess;

Structs

BA512Range

A 512 byte aligned byte range

BlobCacheControl

BlobClient

A client for handling blobs

BlobContentDisposition

BlobContentEncoding

BlobContentLanguage

BlobContentMD5

BlobContentType

BlobLeaseClient

BlobServiceClient

A client for interacting with the blob storage service.

BlockId

Struct wrapping the bytes of a block blob block-id,

CPKInfo

ClientBuilder

A builder for the blob service client.

ConditionAppendPosition

ConditionMaxSize

ContainerClient

ContainerLeaseClient

Cors

CorsRule

CORS is an HTTP feature that enables a web application running under one domain to access resources in another domain. Web browsers implement a security restriction known as same-origin policy that prevents a web page from calling APIs in a different domain; CORS provides a secure way to allow one domain (the origin domain) to call APIs in another domain

EncryptionScope

Logging

Azure Analytics Logging settings.

Metrics

a summary of request statistics grouped by API in hour or minute aggregates for blobs

RetentionPolicy

the retention policy which determines how long the associated data should persist

Snapshot

This type could also be a DateTime but the docs clearly states to treat is as opaque so we do not convert it in any way.

StaticWebsite

The properties that enable an account to host a static website

StoredAccessPolicy

StoredAccessPolicyList

Tags

User-defined tags for specified blobs made up of one or more key-value pairs.

VersionId

This type could also be a DateTime but the docs clearly states to treat is as opaque so we do not convert it in any way.

Enums

AccessTier

BlobExpiry

BlobVersioning

DeleteSnapshotsMethod

Hash

RehydratePriority

Constants

SNAPSHOT