Struct azure_storage_blobs::prelude::BlobClient

pub struct BlobClient { /* private fields */ }

A client for handling blobs

For a full list of operations available on blobs, check out the Azure documentation.

Implementations

impl BlobClient

pub fn from_sas_url(url: &Url) -> Result<Self>

pub fn get(&self) -> GetBlobBuilder

Stream a blob in chunks.

By default, blobs are downloaded in 1MB chunks to reduce the impact of intermittent network issues while downloading large blobs.

pub async fn get_content(&self) -> Result<Vec<u8>>

Return an entire blob.

pub fn get_properties(&self) -> GetPropertiesBuilder

Get all user-defined metadata, standard HTTP properties, and system properties for the blob.

pub fn set_properties(&self) -> SetPropertiesBuilder

Set blob properties.

Several properties are cleared from the blob if not passed. Consider calling set_from_blob_properties with existing blob properties.

pub fn get_metadata(&self) -> GetMetadataBuilder

Get all user-defined metadata for the blob.

pub fn set_metadata(&self) -> SetMetadataBuilder

Set all user-defined metadata of the blob

pub fn set_blob_tier(&self, access_tier: AccessTier) -> SetBlobTierBuilder

Set the access tier on the blob.

pub fn set_blob_expiry(&self, blob_expiry: BlobExpiry) -> SetBlobExpiryBuilder

Set an expiry time on an existing blob.

This operation is only allowed on Hierarchical Namespace enabled accounts.

ref: https://docs.microsoft.com/en-us/rest/api/storageservices/set-blob-expiry

pub fn put_page_blob(&self, length: u128) -> PutPageBlobBuilder

Creates a new page blob.

pub fn put_append_blob(&self) -> PutAppendBlobBuilder

Creates a new append blob.

pub fn put_block_blob(&self, body: impl Into<Body>) -> PutBlockBlobBuilder

Creates a new block blob, or update the content of an existing block blob.

pub fn copy(&self, copy_source: Url) -> CopyBlobBuilder

Copy the blob to a destination within the storage account.

pub fn copy_from_url(&self, copy_source: Url) -> CopyBlobFromUrlBuilder

Copy the blob to a destination within the storage account synchronously.

pub fn acquire_lease<LD: Into<LeaseDuration>>( &self, lease_duration: LD, ) -> AcquireLeaseBuilder

Create a lease on the blob to lock for write and delete operations.

pub fn break_lease(&self) -> BreakLeaseBuilder

End the lease but ensure that another client cannot acquire a new lease until the current lease period has expired.

pub fn delete(&self) -> DeleteBlobBuilder

Delete the blob.

pub fn delete_snapshot(&self, snapshot: Snapshot) -> DeleteBlobSnapshotBuilder

Delete a snapshot of the blob.

pub fn delete_version_id( &self, version_id: VersionId, ) -> DeleteBlobVersionBuilder

Delete the blob at a specific version.

pub fn put_block( &self, block_id: impl Into<BlockId>, body: impl Into<Body>, ) -> PutBlockBuilder

Creates a new block to be committed as part of a block blob.

pub fn put_block_url( &self, block_id: impl Into<BlockId>, copy_source: Url, ) -> PutBlockUrlBuilder

Creates a new block to be committed as part of a block blob, from a URL.

pub fn get_block_list(&self) -> GetBlockListBuilder

Retrieve the list of blocks that have been uploaded as part of a block blob.

pub fn get_tags(&self) -> GetTagsBuilder

Retrieve the user-defined tags for the specified blob, version, or snapshot.

pub fn set_tags(&self, tags: impl Into<Tags>) -> SetTagsBuilder

Set user-defined tags for the specified blob, version, or snapshot.

pub fn put_block_list(&self, block_list: BlockList) -> PutBlockListBuilder

Write a block blob by specifying the list of block IDs that make up the blob.

In order to be written as part of a blob, a block must have been successfully written to the server in a prior Put Block operation.

pub fn put_page( &self, ba512_range: BA512Range, content: impl Into<Body>, ) -> PutPageBuilder

Write a range of pages to a page blob.

pub fn get_page_ranges(&self) -> GetPageRangesBuilder

Return the list of valid page ranges for a page blob or snapshot of a page blob.

pub fn append_block(&self, body: impl Into<Body>) -> AppendBlockBuilder

Commits a new block of data to the end of an existing append blob.

pub fn clear_page(&self, ba512_range: BA512Range) -> ClearPageBuilder

Clear range of pages in a page blob.

pub async fn user_delegation_shared_access_signature( &self, permissions: BlobSasPermissions, user_delegation_key: &UserDeligationKey, ) -> Result<BlobSharedAccessSignature>

pub async fn shared_access_signature( &self, permissions: BlobSasPermissions, expiry: OffsetDateTime, ) -> Result<BlobSharedAccessSignature>

Create a shared access signature.

pub fn generate_signed_blob_url<T>(&self, signature: &T) -> Result<Url>
where T: SasToken,

Create a signed blob url

pub async fn exists(&self) -> Result<bool>

Check whether blob exists.

pub fn snapshot(&self) -> SnapshotBlobBuilder

Create a blob snapshot

pub fn blob_name(&self) -> &str

pub fn blob_lease_client(&self, lease_id: LeaseId) -> BlobLeaseClient

Turn into a BlobLeaseClient

pub fn container_client(&self) -> &ContainerClient

pub fn url(&self) -> Result<Url>

Full URL for the blob.