Struct azure_storage_blobs::prelude::Tags

pub struct Tags {
    pub tag_set: TagSet,
}

User-defined tags for specified blobs made up of one or more key-value pairs.

The following limits apply to blob index tags:

Each blob can have up to 10 tags

Tag keys must be between one and 128 characters

Tag values must be between zero and 256 characters

Tag keys and values are case-sensitive

Tag keys and values only support string data types. Any numbers, dates, times, or special characters are saved as strings

Tag keys and values must adhere to the following naming rules:

Alphanumeric characters:

a through z (lowercase letters)

A through Z (uppercase letters)

0 through 9 (numbers)

Valid special characters: space, plus, minus, period, colon, equals, underscore, forward slash ( +-.:=_/)

ref: https://docs.microsoft.com/en-us/azure/storage/blobs/storage-manage-find-blobs

Fields
tag_set: TagSet
Implementations

impl Tags

pub fn new() -> Self

pub fn insert<K, V>(&mut self, key: K, value: V)
where K: Into<String>, V: Into<String>,

pub fn to_xml(&self) -> Result<Bytes>